Refactor class data and file data into separate container data types
